ISSUE #478 more renaming
